+++ zed_components/src/zed_camera/include/zed_camera_component.hpp
@@ -152,9 +152,6 @@ typedef enum
 // <---- Typedefs to simplify declarations
 
 
-
-
-
 class ZedCamera : public rclcpp::Node
 {
 public:
@@ -404,7 +401,11 @@ private:
   float mTerrainMapPubFreq = 5.0f;  // Frequency of data publishing
   float mTerrainMappingRes = 0.05f;  // Terrain mapping resolution
   float mTerrainMappingRange = 4.0f;  // Terrain mapping range
-  float mTerrainMappingHeigthThresh = 1.0f;  // Max obstacle height
+  float mTerrainMappingRobotHeigth = 1.0f;  // Height of the robot
+  float mTerrainMappingRobotRadius = 0.25f;  // Radius of the robot
+  float mTerrainMappingRobotStep = 0.1f;  // Max height of a step that the robot can overcome
+  float mTerrainMappingRobotSlope = 20.0f;  // Max slope (degrees) that the robot can overcome
+  float mTerrainMappingRobotRoughness = 0.1;  // Max roughness of the terrain that the robot can overcome
   bool mObjDetEnabled = false;
   // TODO(Walter) Add support for Skeleton tracking -> SDK v4!!!
   bool mObjDetTracking = true;
@@ -619,7 +620,7 @@ private:
   // <---- Subscribers
 
   // ----> Terraing Mapping
-  stereolabs::cost_traversability::AgentParameters mAgentParams;
+  stereolabs::cost_traversability::RobotParameters mAgentParams;
   stereolabs::cost_traversability::TraversabilityParameters mTraversabilityParams;
   // <---- Terraing Mapping